// app/types/peopleheader.ts
export interface IndividualHeaderSettings {
    id: number;
    background_gradient_start: string;
    background_gradient_mid: string;
    background_gradient_end: string;
    glow_color: string;
    glow_opacity: number;
    title_prefix: string;
    title_suffix: string;
    description_prefix: string;
    description_italic: string;
    description_suffix: string;
    scroll_arrow_icon_url: string;
    created_at: string;
    updated_at: string;
  }